Require authorization to delete a tag

+++ app/Http/Controllers/TagController.php
@@ -34,6 +34,10 @@ class TagController extends Controller
 
     public function destroy(Tag $tag)
     {
+        if ($tag->user_id !== auth()->id()) {
+            abort(403);
+        }
+
         $tag->delete();
 
         return redirect()->route('tags.index');
